requirements.txt

Django
gunicorn==20.1.0
pytz==2022.4
requests
Pillow
django-sass-processor==1.2.1
libsass
lxml==4.8.0
django-compressor
pycountry==22.3.5
parameterized==0.8.1
celery==5.2.7
redis==4.3.4
django-celery-results==2.4.0
django-celery-beat==2.3.0
psycopg2
black
django-modeltranslation